JCF: Issue #13: add daq_install() function to the DAQ module
A daq_install function, taking a version and a list of targets,
has been added to the DAQ module. The intent is that if a user
wants to install their package in the local development area
(specifically, in the new ./install subdirectory) they can call
this function at the bottom of their CMakeLists.txt file.

In addition to calling this function, they'll also need a file
called <pkgname>Config.cmake.in in the root of their code directory.
They can make one by copying over daq-buildtools' configs/Config.cmake.in
added in this commit, and replacing the comment in the file with the needed
find_dependency() calls.

Once this is done, they can then build and locally install their package
via "build_daq_software.sh --install --pkgname <their package name>"