JCF: make it more obvious when ninja can't get set up
This change has been made because in the last couple of days, not only was
ninja v1_10_1 added to the products area...but ninja v1_8_2 was wiped away.
The consequence of this is that quick-start.sh, which assumed v1_8_2, wasn't
setting up ninja. I've added logic so that if there's a setup failure for
ninja, the scripts print out a warning for you.